First pass at logger
Adding a new logger class that matches most of the Bunyan API for easy
interop.
Moving eureka to a named export instead of a default.
Accept a custom logger via the `logger` option in the constructor.
Reworking some of the APIs to use callbacks.
Consolidating internal cache into one object.